Proposed alternative method for solving star map
I changed the approached to one that didn't use lambdas because I don't think Ben intended for us to be using something complex like that so early on. It is a straightforward code, you have the list count, if the number is odd it becomes * or even this ! so that you get a map of *!*!*! printed out. So yes this is just the pull request thing, sorry if it bothers you, thank you
(lines 57-65)